+++ src/MicroElements.Functional/Messages/MessageExtensions.cs
@@ -20,9 +20,20 @@ namespace MicroElements.Functional
             string eventName = null,
             object state = null,
             IReadOnlyList<KeyValuePair<string, object>> properties = null,
-            IEnumerable<KeyValuePair<string, object>> propertiesEnumerable = null)
+            IEnumerable<KeyValuePair<string, object>> propertiesEnumerable = null,
+            PropertyListAddMode propertyListAddMode = PropertyListAddMode.Set)
         {
-            var propList = properties ?? propertiesEnumerable?.ToList() ?? message.Properties;
+            var propertiesToAdd = properties ?? propertiesEnumerable?.ToList() ?? message.Properties;
+            IReadOnlyList<KeyValuePair<string, object>> propList;
+            if(propertyListAddMode == PropertyListAddMode.Set)
+                propList = propertiesToAdd;
+            else if (propertyListAddMode == PropertyListAddMode.Merge)
+                propList = message.Properties.AddWithReplace(propertiesToAdd);
+            else if (propertyListAddMode == PropertyListAddMode.AddIfNotExists)
+                propList = message.Properties.AddIfNotExists(propertiesToAdd);
+            else
+                propList = Array.Empty<KeyValuePair<string, object>>();
+
             return new Message(
                 timestamp: timestamp ?? message.Timestamp,
                 severity: severity ?? message.Severity,
@@ -77,16 +88,26 @@ namespace MicroElements.Functional
         /// </summary>
         /// <param name="message">Source message.</param>
         /// <param name="properties">New state.</param>
+        /// <param name="propertyListAddMode">Property list add mode.</param>
         /// <returns>New instance of <see cref="Message"/> with changed <see cref="IMessage.Properties"/>.</returns>
-        public static Message WithProperties(this IMessage message, IReadOnlyList<KeyValuePair<string, object>> properties) => message.With(properties: properties);
+        public static Message WithProperties(
+            this IMessage message,
+            IReadOnlyList<KeyValuePair<string, object>> properties,
+            PropertyListAddMode propertyListAddMode = PropertyListAddMode.Set) 
+            => message.With(properties: properties, propertyListAddMode: propertyListAddMode);
 
         /// <summary>
         /// Creates new copy of <see cref="Message"/> with required <see cref="IMessage.Properties"/>.
         /// </summary>
         /// <param name="message">Source message.</param>
         /// <param name="properties">New state.</param>
+        /// <param name="propertyListAddMode">Property list add mode.</param>
         /// <returns>New instance of <see cref="Message"/> with changed <see cref="IMessage.Properties"/>.</returns>
-        public static Message WithProperties(this IMessage message, IEnumerable<KeyValuePair<string, object>> properties) => message.With(propertiesEnumerable: properties);
+        public static Message WithProperties(
+            this IMessage message,
+            IEnumerable<KeyValuePair<string, object>> properties,
+            PropertyListAddMode propertyListAddMode = PropertyListAddMode.Set)
+            => message.With(propertiesEnumerable: properties, propertyListAddMode: propertyListAddMode);
 
         /// <summary>
         /// Creates new copy of <see cref="Message"/> with new property added.
@@ -115,6 +136,52 @@ namespace MicroElements.Functional
         public static Option<object> GetProperty(this IMessage message, string propertyName)
         {
             return message.GetValueAsOption(propertyName);
+        }
+
+        /// <summary>
+        /// Adds KeyValue pairs to source list. Replaces value if exists in source list.
+        /// </summary>
+        /// <param name="source">Source list.</param>
+        /// <param name="keyValuePairs">Items to add.</param>
+        /// <param name="keyEqualityComparer">Key comparer. If not set <see cref="StringComparer.InvariantCultureIgnoreCase"/> will be used.</param>
+        /// <returns>New list with added values.</returns>
+        public static IReadOnlyList<KeyValuePair<string, object>> AddWithReplace(
+            this IEnumerable<KeyValuePair<string, object>> source,
+            IEnumerable<KeyValuePair<string, object>> keyValuePairs,
+            IEqualityComparer<string> keyEqualityComparer = null)
+        {
+            keyEqualityComparer = keyEqualityComparer ?? StringComparer.InvariantCultureIgnoreCase;
+            var dict = source.ToDictionary(pair => pair.Key, pair => pair.Value, keyEqualityComparer);
+            foreach (var valuePair in keyValuePairs)
+            {
+                // Add or replace values
+                dict[valuePair.Key] = valuePair.Value;
+            }
+            return dict.ToList();
+        }
+
+        /// <summary>
+        /// Adds KeysValue pair if not exists in old list.
+        /// </summary>
+        /// <param name="source">Source list.</param>
+        /// <param name="keyValuePairs">Items to add.</param>
+        /// <param name="keyEqualityComparer">Key comparer. If not set <see cref="StringComparer.InvariantCultureIgnoreCase"/> will be used.</param>
+        /// <returns>New list with added values.</returns>
+        public static IReadOnlyList<KeyValuePair<string, object>> AddIfNotExists(
+            this IEnumerable<KeyValuePair<string, object>> source,
+            IEnumerable<KeyValuePair<string, object>> keyValuePairs,
+            IEqualityComparer<string> keyEqualityComparer = null)
+        {
+            keyEqualityComparer = keyEqualityComparer ?? StringComparer.InvariantCultureIgnoreCase;
+            var dict = source.ToDictionary(pair => pair.Key, pair => pair.Value, keyEqualityComparer);
+            foreach (var valuePair in keyValuePairs)
+            {
+                // Skip value is exists in old list.
+                if (dict.ContainsKey(valuePair.Key))
+                    continue;
+                dict.Add(valuePair.Key, valuePair.Value);
+            }
+            return dict.ToList();
         }
     }
 }